+++ fast_track/database/__init__.py
@@ -1,3 +1,0 @@
-""" Initializes the database module. """
-
-from .database import Database